+++ domain/blob_services/blob_fetching_service.py
@@ -1,6 +1,11 @@
 from data.db.db_engine import transactional
 from data.model.blob import Blob
-from data.persistence.blob_reposiotry import get_all_blobs_by_name, get_blob_relative_speeds_by_blob, get_blob_relative_strengths_by_blob
+from data.persistence.blob_reposiotry import (
+    get_all_blobs_by_name,
+    get_blob_by_id,
+    get_blob_relative_speeds_by_blob,
+    get_blob_relative_strengths_by_blob,
+)
 from domain.dtos.blob_stats_dto import BlobStatsDto, IntegrityState, SpeedCategory, StrengthCategory
 from domain.dtos.parent_dto import ParentDto
 from domain.sim_data_service import get_sim_time
@@ -25,7 +30,30 @@ def fetch_all_blobs(
     relative_speeds = get_blob_relative_speeds_by_blob(session)
 
     return [
-        BlobStatsDto(
+        _map_to_blob_state_dto(blob, current_season, relative_speeds.get(blob.id, 0), relative_strengths.get(blob.id, 0))
+        for blob in blobs
+    ]
+
+
+@transactional
+def fetch_blob_by_id(blob_id: int, session) -> BlobStatsDto:
+    """Fetch a blob by its ID and return it as a BlobStatsDto."""
+
+    current_season = get_season(get_sim_time(session))
+
+    relative_strengths = get_blob_relative_strengths_by_blob(session)
+    relative_speeds = get_blob_relative_speeds_by_blob(session)
+
+    blob = get_blob_by_id(session, blob_id)
+
+    if not blob:
+        raise ValueError(f"Blob with ID {blob_id} not found")
+
+    return _map_to_blob_state_dto(blob, current_season, relative_speeds.get(blob.id, 0), relative_strengths.get(blob.id, 0))
+
+
+def _map_to_blob_state_dto(blob: Blob, current_season: int, relative_speed: float, relative_strength: float) -> BlobStatsDto:
+    return BlobStatsDto(
             name=format_blob_name(blob),
             born=format_sim_time_short(blob.born),
             terminated=format_sim_time_short(blob.terminated) if blob.terminated else None,
@@ -45,16 +73,16 @@ def fetch_all_blobs(
             money=blob.money if blob.integrity > 0 else None,
             speed_category=(
                 SpeedCategory.FAST
-                if relative_speeds.get(blob.id, 0) > 0.66
+                if relative_speed > 0.66
                 else SpeedCategory.AVERAGE
-                if relative_speeds.get(blob.id, 0) > 0.33
+                if relative_speed > 0.33
                 else SpeedCategory.SLOW
             ) if blob.integrity > 0 else None,
             strength_category=(
                 StrengthCategory.STRONG
-                if relative_strengths.get(blob.id, 0) > 0.66
+                if relative_strength > 0.66
                 else StrengthCategory.AVERAGE
-                if relative_strengths.get(blob.id, 0) > 0.33
+                if relative_strength > 0.33
                 else StrengthCategory.WEAK
             ) if blob.integrity > 0 else None,
             integrity_state=(
@@ -64,13 +92,11 @@ def fetch_all_blobs(
                 if blob.integrity > 0.4
                 else IntegrityState.POOR
             ) if blob.integrity > 0 else None,
-            speed_color=_get_color_indicator(relative_speeds.get(blob.id, 0)) if blob.integrity > 0 else None,
-            strength_color=_get_color_indicator(relative_strengths.get(blob.id, 0)) if blob.integrity > 0 else None,
+            speed_color=_get_color_indicator(relative_speed) if blob.integrity > 0 else None,
+            strength_color=_get_color_indicator(relative_strength) if blob.integrity > 0 else None,
             integrity_color=_get_color_indicator(blob.integrity / INITIAL_INTEGRITY) if blob.integrity > 0 else None,
             current_activity=blob.current_activity if blob.integrity > 0 and blob.current_activity is not None else None,
         )
-        for blob in blobs
-    ]
 
 
 def _get_color_indicator(relative_value: float) -> str: