+++ CHANGELOG.md
@@ -224,3 +224,4 @@
   - Current activity
 - Added minor blinking animation to blob on blob details dialog
 - Reworked blob activity logic to first proceed with the effects of the current activity stored in the database then generate the activities for next cycle (day)
+- Blob details dialog now can be opened on standings and event pages by clicking on blob name